Remove radio and BSD TV support
BSD actually has support for V4L2, so we don't need the half-baken BSD
specific support code (which probably supports only 1 or 2 old cards).

Radio support is somewhat independent from this, but its configure
checks and even the code are so intermingled with TV and BSD specific
code that we don't bother and remove it as well. Nobody cares.