changes.rst: fuse list of changes from mplayer2 with mpv
There were complaints from Uoti Urpala that the mplayer2 list of changes
was not updated after mplayer2 partially merged some mpv features. This
is a legitimate complaint, so fuse the list of changes for mplayer-svn
to mplayer2 with the list of mpv changes, so that the list of changes is
for mplayer-svn to mpv.